innsending av v7 bassert på toggle (#561)
* ny funksjon for innsending av v7 kontraktVersjon av Søknad. oppretter egen folder for kontrakttyper og refaktorerer

* retter ihht kommentarer

* drar ut nye felter som er lagt til på søknadsobject for å unngå error i loggen for ukjent spørsmålsid under innsending
