+++ src/frontend/components/SøknadsSteg/DinLivssituasjon/NåværendeSamboer.test.tsx
@@ -6,8 +6,8 @@ import { mockDeep } from 'jest-mock-extended';
 import { ESvar } from '@navikt/familie-form-elements';
 
 import { AlternativtSvarForInput } from '../../../typer/common';
-import { ESivilstand } from '../../../typer/person';
-import { ESøknadstype, ISøknad } from '../../../typer/søknad';
+import { ESivilstand, ESøknadstype } from '../../../typer/kontrakt/generelle';
+import { ISøknad } from '../../../typer/søknad';
 import {
     silenceConsoleErrors,
     spyOnUseApp,
