+++ src/frontend/components/SøknadsSteg/DinLivssituasjon/useDinLivssituasjon.tsx
@@ -14,9 +14,10 @@ import useJaNeiSpmFelt from '../../../hooks/useJaNeiSpmFelt';
 import useLanddropdownFeltMedJaNeiAvhengighet from '../../../hooks/useLanddropdownFeltMedJaNeiAvhengighet';
 import { barnDataKeySpørsmål } from '../../../typer/barn';
 import { AlternativtSvarForInput } from '../../../typer/common';
-import { Dokumentasjonsbehov } from '../../../typer/dokumentasjon';
+import { Dokumentasjonsbehov } from '../../../typer/kontrakt/dokumentasjon';
+import { ESivilstand } from '../../../typer/kontrakt/generelle';
 import { IArbeidsperiode, IPensjonsperiode } from '../../../typer/perioder';
-import { ESivilstand, ISamboer, ISøker, ITidligereSamboer } from '../../../typer/person';
+import { ISamboer, ISøker, ITidligereSamboer } from '../../../typer/person';
 import { IDinLivssituasjonFeltTyper } from '../../../typer/skjema';
 import { Årsak } from '../../../typer/utvidet';
 import { dagensDato } from '../../../utils/dato';
