+++ src/frontend/components/SøknadsSteg/OmBarnaDine/useOmBarnaDine.tsx
@@ -7,8 +7,8 @@ import { useApp } from '../../../context/AppContext';
 import { useEøs } from '../../../context/EøsContext';
 import useJaNeiSpmFelt from '../../../hooks/useJaNeiSpmFelt';
 import { barnDataKeySpørsmål } from '../../../typer/barn';
-import { Dokumentasjonsbehov } from '../../../typer/dokumentasjon';
-import { ESivilstand } from '../../../typer/person';
+import { Dokumentasjonsbehov } from '../../../typer/kontrakt/dokumentasjon';
+import { ESivilstand } from '../../../typer/kontrakt/generelle';
 import { IOmBarnaDineFeltTyper } from '../../../typer/skjema';
 import { Årsak } from '../../../typer/utvidet';
 import { genererOppdaterteBarn } from '../../../utils/barn';
