+++ src/frontend/components/SøknadsSteg/Oppsummering/OppsummeringSteg/DinLivssituasjonOppsummering.tsx
@@ -11,7 +11,8 @@ import { useApp } from '../../../../context/AppContext';
 import { useFeatureToggles } from '../../../../context/FeatureToggleContext';
 import { useRoutes } from '../../../../context/RoutesContext';
 import { AlternativtSvarForInput } from '../../../../typer/common';
-import { ESivilstand, ISamboer, ITidligereSamboer } from '../../../../typer/person';
+import { ESivilstand } from '../../../../typer/kontrakt/generelle';
+import { ISamboer, ITidligereSamboer } from '../../../../typer/person';
 import { RouteEnum } from '../../../../typer/routes';
 import { formaterDato } from '../../../../utils/dato';
 import { landkodeTilSpråk, toÅrsakSpråkId } from '../../../../utils/språk';
