+++ src/frontend/context/EøsContext.ts
@@ -9,7 +9,8 @@ import { RessursStatus } from '@navikt/familie-typer';
 import Miljø from '../Miljø';
 import { andreForelderDataKeySpørsmål, barnDataKeySpørsmål, IBarnMedISøknad } from '../typer/barn';
 import { BarnetsId } from '../typer/common';
-import { Dokumentasjonsbehov, IDokumentasjon } from '../typer/dokumentasjon';
+import { IDokumentasjon } from '../typer/dokumentasjon';
+import { Dokumentasjonsbehov } from '../typer/kontrakt/dokumentasjon';
 import { ISøker } from '../typer/person';
 import { useApp } from './AppContext';
 import { useFeatureToggles } from './FeatureToggleContext';
