+++ src/frontend/typer/dokumentasjon.ts
@@ -1,16 +1,4 @@
-import { LocaleType } from '@navikt/familie-sprakvelger';
-
-export enum Dokumentasjonsbehov {
-    AVTALE_DELT_BOSTED = 'AVTALE_DELT_BOSTED',
-    VEDTAK_OPPHOLDSTILLATELSE = 'VEDTAK_OPPHOLDSTILLATELSE',
-    ADOPSJON_DATO = 'ADOPSJON_DATO',
-    BEKREFTELSE_FRA_BARNEVERN = 'BEKREFTELSE_FRA_BARNEVERN',
-    BOR_FAST_MED_SØKER = 'BOR_FAST_MED_SØKER',
-    SEPARERT_SKILT_ENKE = 'SEPARERT_SKILT_ENKE',
-    MEKLINGSATTEST = 'MEKLINGSATTEST',
-    EØS_SKJEMA = 'EØS_SKJEMA',
-    ANNEN_DOKUMENTASJON = 'ANNEN_DOKUMENTASJON',
-}
+import { Dokumentasjonsbehov } from './kontrakt/dokumentasjon';
 
 export interface IVedlegg {
     dokumentId: string;
@@ -27,19 +15,6 @@ export interface IDokumentasjon {
     opplastedeVedlegg: IVedlegg[];
     tittelSpråkId: string;
     beskrivelseSpråkId: string | null;
-}
-
-export interface ISøknadKontraktVedlegg {
-    dokumentId: string;
-    navn: string;
-    tittel: Dokumentasjonsbehov;
-}
-
-export interface ISøknadKontraktDokumentasjon {
-    dokumentasjonsbehov: Dokumentasjonsbehov;
-    harSendtInn: boolean;
-    opplastedeVedlegg: ISøknadKontraktVedlegg[];
-    dokumentasjonSpråkTittel: Record<LocaleType, string>;
 }
 
 export enum EFiltyper {
