+++ src/frontend/typer/kontrakt/barn.ts
@@ -1,0 +1,17 @@
+import {
+    ERegistrertBostedType,
+    IAndreForelderIKontraktFormat,
+    ISøknadsfelt,
+    IUtenlandsperiodeIKontraktFormat,
+    SpørsmålMap,
+} from './generelle';
+
+export interface ISøknadKontraktBarn {
+    ident: ISøknadsfelt<string>;
+    navn: ISøknadsfelt<string>;
+    registrertBostedType: ISøknadsfelt<ERegistrertBostedType>;
+    alder: ISøknadsfelt<string>;
+    spørsmål: SpørsmålMap;
+    utenlandsperioder: ISøknadsfelt<IUtenlandsperiodeIKontraktFormat>[];
+    andreForelder: IAndreForelderIKontraktFormat | null;
+}
