+++ src/frontend/utils/språk.test.ts
@@ -1,6 +1,6 @@
 import { LocaleType } from '@navikt/familie-sprakvelger';
 
-import { ESivilstand } from '../typer/person';
+import { ESivilstand } from '../typer/kontrakt/generelle';
 import { hentSivilstatusSpråkId, hentTekster, landkodeTilSpråk } from './språk';
 
 describe('hentTekster', () => {
