+++ src/frontend/utils/testing.tsx
@@ -34,9 +34,10 @@ import { StegProvider } from '../context/StegContext';
 import { andreForelderDataKeySpørsmål, barnDataKeySpørsmål } from '../typer/barn';
 import { AlternativtSvarForInput } from '../typer/common';
 import { EFeatureToggle } from '../typer/feature-toggles';
+import { ESivilstand, ESøknadstype } from '../typer/kontrakt/generelle';
 import { IKvittering } from '../typer/kvittering';
-import { ESivilstand, ISøker, ISøkerRespons } from '../typer/person';
-import { ESøknadstype, initialStateSøknad, ISøknad } from '../typer/søknad';
+import { ISøker, ISøkerRespons } from '../typer/person';
+import { initialStateSøknad, ISøknad } from '../typer/søknad';
 import { Årsak } from '../typer/utvidet';
 import { genererInitialBarnMedISøknad } from './barn';
 import * as eøsUtils from './eøs';
